backend/pyproject.toml
[project]
name = "dhandho"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "beautifulsoup4>=4.14.2",
    "fastapi>=0.118.2",
    "groq>=0.32.0",
    "lxml>=6.0.2",
    "pydantic>=2.12.0",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn>=0.37.0",
]
